// src/app/core/models/regularOrder.model.ts
export class regularOrderModel extends PaginationModel {

  constructor() {
    super();
    this.SurpressEmailNotification = false;
  }
  orderWorkbenchFilterModal = new OrderWorkbenchFilterModal();
  OrderType: any;
  OrganizationID: number;
  ClientId: number;
  UserType: string;
  OrgCode: string;
  OrderStatus: number;
  ShipTo: number;
  ShipFrom: number;
  SurpressEmailNotification: boolean;
  SalesManager: string;
  purchaseOrderNumber: string;
  AvailableCredit: string;
  invoiceNo: string;
  allocatedFreightCharge: number;
  ajustmentChagresMaterialID: number;
  ajustmentChagresChargeID: number;
  shippingToContractID: number;
  LocationType: string;
  RoleName: string;
  InboundOutbound: string;
  UserLoginID: string;
}

export class PeriodicElement {
  Id: number;
  LoginId: string;
  UserName: string;
  CodeMobilePhone: string;
  OrganizationName: string;
  RoleName: string;
  LocationName: string;
  Status: string;
  IsSeleted: boolean = false;
  constructor() {
    this.IsSeleted = false;
  }

}

export class CommonOrderModel {
  OrderTypeId: number;
  ClientID: number;
  LocationFunctionID: number;
  LocationID: number;
  Action: string;
}
// ...
export class OrderWorkbenchFilterModal {
  OrderNo: string;
  selectedOrderTypeItems = [];
  selectedOrderStatusItems = [];
  selectedOrderConditionItems = [];
  MaxOrderAmount: number;
  MinOrderAmount: number;
  selectedShipFromItems = [];
  selectedShipToItems = [];
  selectedShipFromTypeItems = [];
  selectedShipToTypeItems = [];
  ShipmentNo: string;
  PONo: string;
  selectedMaterialItems = [];
  selectedCarrierItems = [];
  isShipWith: boolean;
  ScheduledShipDateN: string;
  ReqDeliveryDateN: string;
  SourceN: string;
  selectedSalesManagersItems = [];
  selectedInvoiceNoItems = [];
  selectedAllVersionItems = [];
  selectedSourceItems = [];
  MustArriveByDateN: string;
  ShipWithOption:string


}